Changed Kernel to actual kernel in OS X
It was previously gathed from "uname" which returned darwin. Darwin is
not the kernel but rather the operating system OS X is based on. Since
this application is designed only to run on OS X, and all versions of
OS X run on XNU. So the kernel variable can just be a string.