Updated version code
Had a good solution for bash 4.0, but as mention does not come by
default in OS X, and bash 3 solution works for both so only left that
in. Changed variable names to make a bit more readable. Fixed a bug for
releases such as “10.9.1” that caused a parse error because the code
was trying to multiply a number with 2 decimal places.